Fixed App::import of Yaml.php again :-P (read on)
App::import('Vendor', 'Migrations.Yaml');
only works if db-migrations is installed as a plugin on a
case-insensitive filesystem. The last part is important, as the import
script searches for "yaml.php" and not for "Yaml.php", which is equal on
a case-insensitive filesystem.

Adding the file parameter to the import method makes cake look for the
right file and interestingly also makes it search in every place
possible (even in every plugin).

So hopefully this works.